+++ packages/expo-router/src/global-state/routing.ts
@@ -1,4 +1,4 @@
-import { StackActions, type NavigationState } from '@react-navigation/native';
+import { StackActions, type NavigationState, PartialRoute } from '@react-navigation/native';
 import * as Linking from 'expo-linking';
 import { nanoid } from 'nanoid/non-secure';
 
@@ -7,6 +7,7 @@ import { ExpoRouter } from '../../types/expo-router';
 import { ResultState } from '../fork/getStateFromPath';
 import { resolveHref } from '../link/href';
 import { resolve } from '../link/path';
+import { matchDynamicName } from '../matchers';
 import { shouldLinkExternally } from '../utils/url';
 
 function assertIsReady(store: RouterStore) {
@@ -141,36 +142,78 @@ export function linkTo(this: RouterStore, href: string, event?: string) {
   return navigationRef.dispatch(getNavigateAction(state, rootState, event));
 }
 
-type NavigationParams = Partial<{
-  screen: string;
-  params: NavigationParams;
-  key?: string;
-}>;
-
-function rewriteNavigationStateToParams(
-  state: { routes: ResultState['routes'] } | NavigationState | undefined,
-  params: NavigationParams = {}
+function getNavigateAction(
+  actionState: ResultState,
+  navigationState: NavigationState,
+  type = 'NAVIGATE'
 ) {
-  if (!state) return params;
-  // We Should always have at least one route in the state
-  const lastRoute = state.routes[state.routes.length - 1]!;
-  params.screen = lastRoute.name;
-  // Weirdly, this always needs to be an object. If it's undefined, it won't work.
-  params.params = lastRoute.params ? JSON.parse(JSON.stringify(lastRoute.params)) : {};
+  /**
+   * We need to find the deepest navigator where the action and current state diverge, If they do not diverge, the
+   * lowest navigator is the target.
+   *
+   * By default React Navigation will target the current navigator, but this doesn't work for all actions
+   * For example:
+   *  - /deeply/nested/route -> /top-level-route the target needs to be the top-level navigator
+   *  - /stack/nestedStack/page -> /stack1/nestedStack/other-page needs to target the nestedStack navigator
+   *
+   * This matching needs to done by comparing the route names and the dynamic path, for example
+   * - /1/page -> /2/anotherPage needs to target the /[id] navigator
+   *
+   * Other parameters such as search params and hash are not evaluated.
+   *
+   */
+  let actionStateRoute: PartialRoute<any> | undefined;
 
-  if (lastRoute.state) {
-    rewriteNavigationStateToParams(lastRoute.state, params.params);
+  // Traverse the state tree comparing the current state and the action state until we find where they diverge
+  while (actionState && navigationState) {
+    const stateRoute = navigationState.routes[navigationState.index];
+
+    actionStateRoute = actionState.routes[actionState.routes.length - 1];
+
+    const childState = actionStateRoute.state;
+    const nextNavigationState = stateRoute.state;
+
+    const dynamicName = matchDynamicName(actionStateRoute.name);
+
+    const didActionAndCurrentStateDiverge =
+      actionStateRoute.name !== stateRoute.name ||
+      !childState ||
+      !nextNavigationState ||
+      (dynamicName && actionStateRoute.params?.[dynamicName] !== stateRoute.params?.[dynamicName]);
+
+    if (didActionAndCurrentStateDiverge) {
+      break;
+    }
+
+    actionState = childState;
+    navigationState = nextNavigationState as NavigationState;
   }
 
-  return JSON.parse(JSON.stringify(params));
-}
+  /*
+   * We found the target navigator, but the payload is in the incorrect format
+   * We need to convert the action state to a payload that can be dispatched
+   */
+  const rootPayload: Record<string, any> = { params: {} };
+  let payload = rootPayload;
+  let params = payload.params;
 
-function getNavigateAction(state: ResultState, parentState: NavigationState, type = 'NAVIGATE') {
-  const { screen, params } = rewriteNavigationStateToParams(state);
+  // The root level of payload is a bit weird, its params are in the child object
+  while (actionStateRoute) {
+    Object.assign(params, { ...actionStateRoute.params });
+    payload.screen = actionStateRoute.name;
 
-  let key: string | undefined;
+    actionStateRoute = actionStateRoute.state?.routes[actionStateRoute.state?.routes.length - 1];
 
+    payload.params ??= {};
+    payload = payload.params;
+    params = payload;
+  }
+
+  // Expo Router uses only three actions, but these don't directly translate to all navigator actions
   if (type === 'PUSH') {
+    // Only stack navigators have a push action, and even then we want to use NAVIGATE (see below)
+    type = 'NAVIGATE';
+
     /*
      * The StackAction.PUSH does not work correctly with Expo Router.
      *
@@ -185,22 +228,22 @@ function getNavigateAction(state: ResultState, parentState: NavigationState, typ
      * By generating a unique new key, we ensure that the screen is always pushed onto the stack.
      *
      */
-    type = 'NAVIGATE';
-
-    if (parentState.type === 'stack') {
-      key = `${screen}-${nanoid()}`; // @see https://github.com/react-navigation/react-navigation/blob/13d4aa270b301faf07960b4cd861ffc91e9b2c46/packages/routers/src/StackRouter.tsx#L406-L407
+    if (navigationState.type === 'stack') {
+      rootPayload.key = `${rootPayload.name}-${nanoid()}`; // @see https://github.com/react-navigation/react-navigation/blob/13d4aa270b301faf07960b4cd861ffc91e9b2c46/packages/routers/src/StackRouter.tsx#L406-L407
     }
-  } else if (type === 'REPLACE' && parentState.type === 'tab') {
+  }
+
+  if (type === 'REPLACE' && navigationState.type === 'tab') {
     type = 'JUMP_TO';
   }
 
   return {
     type,
-    target: parentState.key,
+    target: navigationState.key,
     payload: {
-      key,
-      name: screen,
-      params,
+      key: rootPayload.key,
+      name: rootPayload.screen,
+      params: rootPayload.params,
     },
   };
 }